Move constraint instantiation to appropriate file
string_constraint_instantiation.cpp is the appropriate place for the
instanciation functions.
